Create MarbleCollector.java
Marble Collector
There are a total of 
M
M possible types of marbles that you want to collect. They are numbered 
1
,
2
,
…
,
M
1,2,…,M.

You currently have 
N
N marbles with you, with types 
A
1
,
A
2
,
…
,
A
N
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
 . Note that you may have the same type of marble twice or more.

Find number of types of marbles that you still have not collected.

Input Format
The first line of input will contain a single integer 
T
T, denoting the number of test cases.
Each test case consists of multiple lines of input.
The first line of each test case contains 
N
N and 
M
M - the number of marbles you have, and the total possible types of marbles.
The second line contains 
N
N integers - 
A
1
,
A
2
,
…
,
A
N
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
 .
Output Format
For each test case, output on a new line the number of types of marbles you have not collected yet.